Feature ETP-456: Fix cache store and circular dependencies.
