+++ packages/MainUI/src/utils/metadata.ts
@@ -1,7 +1,4 @@
-import {
-  type Etendo,
-  Metadata,
-} from '@workspaceui/etendohookbinder/api/metadata';
+import { type Etendo, Metadata } from '@workspaceui/etendohookbinder/api/metadata';
 
 export const groupTabsByLevel = (windowData?: Etendo.WindowMetadata) => {
   if (!windowData) {
@@ -35,4 +32,32 @@ export const buildColumnsData = (windowData?: Etendo.WindowMetadata) => {
   }
 
   return cols;
+};
+
+export const parseColumns = (columns?: Etendo.Field[]): Etendo.Column[] => {
+  const result: Etendo.Column[] = [];
+
+  if (!columns) return result;
+
+  for (const column of columns) {
+    if (column.showInGridView) {
+      result.push({
+        header: column.title ?? column.name ?? column.columnName,
+        id: column.name,
+        columnName: column.columnName,
+        isMandatory: column.required,
+        _identifier: column.title,
+        column: {
+          _identifier: column.title,
+          reference: column.type,
+        },
+        name: column.name,
+        accessorFn: (v: Record<string, unknown>) => {
+          return v[column.columnName + '$_identifier'] ?? v[column.columnName];
+        },
+      });
+    }
+  }
+
+  return result;
 };
